// new_dashboard/types/GameData.ts
type GameData = {
    createdDate: string;
    creator: string;
    favoritedCount: string;
    genre: string;
    link: string;
    placeId: string;
    playing: string;
    rating: string;
    title: string;
    universeId: string;
    visits: string;
    upvote: string;
    downvote: string;
    imageUrl: string;
}

export type { GameData };